Consolidate `refine-fixtures` and `refine-types` to `refine-core`
