+++ package.json
@@ -5,7 +5,7 @@
     "packages/*"
   ],
   "scripts": {
-    "dev:react": "concurrently \"yarn workspace refine-types dev\" \"yarn workspace refine-fixtures dev\" \"yarn workspace @hammerstone/refine-react dev\" \"yarn workspace react dev\"",
+    "dev:react": "concurrently \"yarn workspace refine-core dev\" \"yarn workspace @hammerstone/refine-react dev\" \"yarn workspace react dev\"",
     "dev:vue2": "concurrently \"yarn workspace @hammerstone/refine-vue build:vue2 --watch\" \"yarn workspace vue2 dev --watch\" ",
     "dev:vue3": "yarn workspace vue3 dev"
   },
