+++ packages/core/.gitignore
