+++ packages/core/src/types/conditions.ts
