+++ packages/core/tsconfig.json
@@ -1,0 +1,6 @@
+{
+  "compilerOptions": {
+    "moduleResolution": "node",
+    "baseUrl": "./src"
+  }
+}
