--- packages/fixtures/tsconfig.json
@@ -1,5 +1,0 @@
-{
-  "compilerOptions": {
-    "moduleResolution": "node"
-  }
-}
