+++ packages/refine-react/package.json
@@ -26,8 +26,7 @@
     "@types/jest": "^27.4.0",
     "@types/react": "^17.0.0",
     "jest": "^27.4.7",
-    "refine-fixtures": "*",
-    "refine-types": "*",
+    "refine-core": "*",
     "tsup": "^5.11.11",
     "typescript": "^4.5.5"
   }
