+++ packages/refine-react/src/tailwind/components/criterion-group/use-criterion-group.ts
@@ -1,5 +1,5 @@
 import { createContext, useContext } from "react";
-import { CriterionBlueprintItem } from "refine-types";
+import { CriterionBlueprintItem } from "refine-core/types";
 
 export type CriterionGroupContext = {
   index: number;
