+++ packages/refine-react/src/tailwind/components/inputs/double-date/double-date-input.test.tsx
@@ -1,5 +1,5 @@
 import { render, screen } from "@testing-library/react";
-import { dateCondition } from "refine-fixtures";
+import { dateCondition } from "refine-core/fixtures";
 import { QueryBuilder } from "../..";
 
 it("should render two date inputs", async () => {
