--- packages/types/.gitignore
@@ -1,1 +1,0 @@
-dist
