--- packages/types/package.json
@@ -1,21 +1,0 @@
-{
-  "name": "refine-types",
-  "version": "0.0.0",
-  "main": "./dist/index.js",
-  "module": "./dist/index.mjs",
-  "types": "./dist/index.d.ts",
-  "sideEffects": false,
-  "license": "MIT",
-  "files": [
-    "dist/**"
-  ],
-  "packageManager": "yarn@3.1.1",
-  "scripts": {
-    "build": "tsup src/index.ts --format esm,cjs --dts",
-    "dev": "tsup src/index.ts --format esm,cjs --watch --dts"
-  },
-  "devDependencies": {
-    "tsup": "^5.11.11",
-    "typescript": "^4.5.5"
-  }
-}
